将req用pb转为native层的AutoBuffer

+++ app/src/main/cpp/network/autobuffer.h
@@ -11,13 +11,13 @@ class AutoBuffer {
 
     void Write(const unsigned char *_byte_array, size_t _len);
 
-    off_t Pos() const;
+    size_t Pos() const;
 
     size_t Length() const;
 
     void SetLength(size_t _len);
 
-    unsigned char *Ptr(const off_t _offset = 0) const;
+    unsigned char *Ptr(const size_t _offset = 0) const;
 
     size_t GetCapacity() const;
 
@@ -28,7 +28,7 @@ class AutoBuffer {
 
   private:
     unsigned char * byte_array_;
-    off_t           pos_;
+    size_t           pos_;
     size_t          length_;
     size_t          capacity_;
     const size_t    malloc_unit_size_;
